requirements.txt
flask
tensorflow
pillow
pandas
numpy
matplotlib
scikit-learn
transformers
sentence-transformers
accelerate
torch
huggingface-hub
